requirements.txt
# Skin Assistant API & ML
pandas>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
openai>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0

# Optional: Streamlit chat UI
# streamlit>=1.28.0

# Optional: product image model training (PyTorch)
# torch>=2.0.0
# torchvision>=0.15.0
